moved roleNames from userRoles into UserList component
