+++ src/client/components/UserList/UserList.tsx
@@ -2,11 +2,19 @@ import './UserList.scss'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { User } from '@meta/user'
+import { RoleName, User } from '@meta/user'
 
 import AdministrationListElement from './AdministrationListElement'
 import CollaboratorListElement from './CollaboratorListElement'
 import UserListHeader from './UserListHeader'
+
+const roleNames = [
+  RoleName.REVIEWER,
+  RoleName.NATIONAL_CORRESPONDENT,
+  RoleName.ALTERNATE_NATIONAL_CORRESPONDENT,
+  RoleName.COLLABORATOR,
+  RoleName.VIEWER,
+]
 
 type Props = {
   users: Array<User>
@@ -18,11 +26,11 @@ const UserList: React.FC<Props> = ({ users, isAdmin }) => {
 
   return users && users.length > 0 ? (
     <table className="user-list__table">
-      <UserListHeader isAdmin={isAdmin} />
+      <UserListHeader isAdmin={isAdmin} roleNames={roleNames} />
       <tbody>
         {users.map((user: User) =>
           isAdmin ? (
-            <AdministrationListElement key={user.id} user={user} />
+            <AdministrationListElement key={user.id} user={user} roleNames={roleNames} />
           ) : (
             <CollaboratorListElement key={user.id} user={user} />
           )
